refactor: don't use `unemojify()`
This wasn't actually being used like I thought it was, huge thanks to
@cinnamon-bun for taking the time to point this out! Yay, fewer moving
parts.

Warning: super similar code in lib/block and lib/inline that should
probably be refactored in the future. Not similar enough that it's easy
to refactor right now. (Read: energy dwindling at the expense of best
practices.)