Fix external images being rendered by default
Problem: We don't want external HTTP requests to be made so it's bad if
people can embed images with remote URLS. This library was allowing that
behavior to happen by default, although clients could fix this with the
`imageLink()` method.

Solution: Always check Markdown images (which includes audio and video
(!)) and strip all values that aren't valid blob references according to
SSB-Ref.